Type subsystem
- Awesome error messages for unkown types.
- Tracks types for variables.

+++ tests/type1.pas
@@ -1,6 +1,6 @@
 program type;
 type
-  TAngle = byte;
+  TAngle = Integer;
 var
   alpha: TAngle;
 begin
